requirements.txt
yadisk>=1.2.0,<2.0.0
webdavclient3>=3.14.0
requests>=2.28.0
PyQt6>=6.5.0
Pillow>=10.0.0
watchdog>=6.0.0